feat: Implement refresh tokens
Addresses #160

Access tokens expire and require the user to login again
in order to receive a new one. With the use of refresh
tokens, the user can request from the auth server new
access tokens without having to logout and login again.
Moreover, the life of a refresh token is much longer, but
for this reason they need to be stored very securely.

Signed-off-by: Iuliana Voinea <iulianavoinea96@gmail.com>