Fixed, searching based on both the username and name #745
Now, the list_users_api returns the results on present of both
username and name in the search_query.

To solve this issue added a OR operator to search on basis of
username as well as name.

Tests: Added unittest for this in test_api_list_users.py named as
test_list_users_api_with_a_search_query_with_username_resource_auth

Fixes #745